Refactored FindDI_Ready into multiple functions
In order to continue the automation process, it was necessary to
separate some of the logic into independent functions.

The original functionality of the script has been preserved --
that is, a single archive or volume may be specified, and the
program will return a message regarding current/expired status.

Additionally, this change allows for a more descriptive error to
be printed in the case that the database connection fails.