Refactored db_connect to return a bound session and exclude the return of mapped classes.
db_connect previously returned an unbound session that required an explicite session(bind=engine)
every time.  Since this logic is reused and required, it makes sense for it to be in the function.

Additionally, this commit factors out a temporary fix involving the return of Files and Archives
classes that are mapped using metadata.  This should be a lot cleaner.

It may not be necessary to return the engine - it seems to be the case that we're ignoring / "wild-carding"
the output in the vast majority of cases.  I'm not sure when the engine is necessary except for binding, but
I'm leaving it in for the time being.