Initial xml2json.py
Needs to be made more general.  Initial version was based on the
mroctx json recipes, so instrument specific information may be
missing when converting other recipes.

Some information is missing, some is scattered, and some is found
in different areas among xml recipes.  This is not yet a general
purpose tool, but I'm working on improvements that will make it more
general.